+++ .formatter.exs
@@ -2,7 +2,8 @@ locals_without_parens = [
   defvocab: 2,
   def_facet_constraint: 2,
   def_applicable_facet: 1,
-  bgp: 1
+  bgp: 1,
+  build: 2
 ]
 
 [